Move helper functions into dedicated module
