+++ diffsky/data_loaders/opencosmo_utils/compute.py
@@ -6,7 +6,6 @@ import numpy as np
 import opencosmo as oc
 from diffmah import DiffmahParams
 from diffstar import DiffstarParams
-from dsps.cosmology import age_at_z
 
 from ... import phot_utils
 from ...experimental import precompute_ssp_phot as psspp
@@ -17,31 +16,7 @@ from ...experimental.kernels import (
     phot_kernels_merging,
     sed_kernels_merging,
 )
-
-
-def __get_z_phot_tables(catalog: oc.Lightcone):
-    """
-    Retrieve z_phot_tables from the underlying catalogs, or construct
-    them if not found.
-    """
-    z_phot_tables = {}
-    for slice_name, dataset in catalog.items():
-        z_phot_tables[slice_name] = __get_z_phot_table_from_dataset(dataset)
-    return z_phot_tables
-
-
-def __get_z_phot_table_from_dataset(dataset: oc.Dataset):
-    z_phot_table = dataset.header.catalog_info.get("zphot_table")
-    if z_phot_table is None:
-        return __estimate_z_phot_table(dataset)
-    return np.array(z_phot_table)
-
-
-def __estimate_z_phot_table(dataset: oc.Dataset):
-    min_z, max_z = dataset.header.lightcone["z_range"]
-    min_z = 0.95 * min_z
-    max_z = 1.05 * max_z
-    return np.linspace(min_z, max_z, 15)
+from . import utils
 
 
 def compute_phot_from_diffsky_mock(
@@ -90,16 +65,16 @@ def compute_phot_from_diffsky_mock(
 
 
     """
-    __validate_batch_size(batch_size)
+    utils.validate_batch_size(batch_size)
     func = phot_kernels_merging._phot_kern_merging
-    z_phot_tables = __get_z_phot_tables(catalog)
+    z_phot_tables = utils.get_z_phot_tables(catalog)
     suffix = ""
     if set(bands).intersection(catalog.columns):
         suffix = "_new"
 
     result = __run_photometry(
         func,
-        __unpack_photometry,
+        utils.unpack_photometry,
         catalog,
         aux_data,
         z_phot_tables,
@@ -160,16 +135,16 @@ def compute_dbk_phot_from_diffsky_mock(
         insert = False, return the computed photometry directly.
 
     """
-    __validate_batch_size(batch_size)
+    utils.validate_batch_size(batch_size)
 
     suffix = ""
     if set(bands).intersection(catalog.columns):
         suffix = "_new"
-    z_phot_tables = __get_z_phot_tables(catalog)
+    z_phot_tables = utils.get_z_phot_tables(catalog)
     func = dbk_specphot_kernels_merging._dbk_specphot_kern_merging
     result = __run_photometry(
         func,
-        __unpack_dbk_photometry,
+        utils.unpack_dbk_photometry,
         catalog,
         aux_data,
         z_phot_tables,
@@ -222,7 +197,7 @@ def compute_seds_from_diffsky_mock(
         insert = False, return the computed photometry directly.
 
     """
-    __validate_batch_size(batch_size)
+    utils.validate_batch_size(batch_size)
 
     result = __run_sed_computation(catalog, aux_data, __compute_sed_managed, batch_size)
     if insert:
@@ -269,63 +244,13 @@ def compute_dbk_seds_from_diffsky_mock(
 
 
     """
-    __validate_batch_size(batch_size)
+    utils.validate_batch_size(batch_size)
     result = __run_sed_computation(
         catalog, aux_data, __compute_dbk_sed_managed, batch_size
     )
     if insert:
         return catalog.with_new_columns(**result)
     return result
-
-
-def __unpack_photometry(data, band_names, suffix, gal_id, *args):
-    result = __unpack_photometry_array(data.obs_mags, band_names, suffix)
-    result["gal_id"] = gal_id
-    return result
-
-
-def __unpack_dbk_photometry(data, band_names, suffix, gal_id, include_extras):
-
-    phot_info, dbk_weights = data
-    obs_mags_disk = phot_info.obs_mags_disk
-    obs_mags_bulge = phot_info.obs_mags_bulge
-    obs_mags_knot = phot_info.obs_mags_knots
-
-    bulge_bands = [f"{bn}_bulge" for bn in band_names]
-    disk_bands = [f"{bn}_disk" for bn in band_names]
-    knot_bands = [f"{bn}_knots" for bn in band_names]
-
-    output = __unpack_photometry_array(obs_mags_bulge, bulge_bands, suffix)
-    output |= __unpack_photometry_array(obs_mags_disk, disk_bands, suffix)
-    output |= __unpack_photometry_array(obs_mags_knot, knot_bands, suffix)
-    output["gal_id"] = gal_id
-    if include_extras is not None:
-        phot_info = phot_info._asdict()
-        output |= {name: np.array(phot_info[name]) for name in include_extras}
-
-    return output
-
-
-def __unpack_photometry_array(data, band_names, suffix):
-    to_unpack = np.array(data).T
-    return {f"{name}{suffix}": to_unpack[i] for i, name in enumerate(band_names)}
-
-
-def __validate_batch_size(batch_size):
-    if batch_size == -1:
-        return
-    if not isinstance(batch_size, int) or batch_size < 1:
-        raise ValueError(
-            f"batch_size must be -1 (no batching) or a positive integer, got {batch_size!r}"
-        )
-
-
-def __split_central_indices(catalog: oc.Lightcone, batch_size: int):
-    central_indices = np.where(catalog.select("central").get_data("numpy"))[0]
-    if batch_size == -1:
-        return [central_indices]
-    split_indices = np.arange(batch_size, len(central_indices), batch_size)
-    return np.split(central_indices, split_indices)
 
 
 def __run_photometry(
@@ -355,7 +280,7 @@ def __run_photometry(
 
     wave_eff_tables = {}
     precomputed_ssp_mag_tables = {}
-    cosmology_parameters = __prep_cosmology_parameters(catalog.cosmology)
+    cosmology_parameters = utils.prep_cosmology_parameters(catalog.cosmology)
 
     for slice_name, z_phot_table in z_phot_tables.items():
         wave_eff_tables[slice_name] = phot_utils.get_wave_eff_table(
@@ -367,7 +292,7 @@ def __run_photometry(
             )
         )
     catalog = catalog.evaluate(
-        age_at_z_,
+        utils.age_at_z_,
         vectorize=True,
         cosmology=cosmology_parameters,
         format="jax",
@@ -377,7 +302,7 @@ def __run_photometry(
     else:
         to_compute = __compute_photometry_managed
 
-    batches = __split_central_indices(catalog, batch_size)
+    batches = utils.split_central_indices(catalog, batch_size)
 
     chunked_output = []
     for row_batch in batches:
@@ -426,43 +351,21 @@ def __run_photometry(
     return {name: band[permutation] for name, band in output.items()}
 
 
-def __prep_cosmology_parameters(cosmology):
-    try:
-        w0 = cosmology.wo
-        wa = cosmology.wa
-    except AttributeError:  # Why astropy... Why...
-        w0 = -1
-        wa = 0
-    Cosmology_t = namedtuple("Cosmology_t", ("Om0", "w0", "wa", "h"))
-    return Cosmology_t(cosmology.Om0, w0, wa, cosmology.h)
-
-
-def age_at_z_(redshift_true, cosmology):
-    result = {
-        "t_obs": np.array(
-            age_at_z(
-                redshift_true, cosmology.Om0, cosmology.w0, cosmology.wa, cosmology.h
-            )
-        )
-    }
-    return result
-
-
 def __run_sed_computation(
     catalog: oc.Lightcone,
     aux_data: dict,
     to_compute: Callable,
     batch_size: int = 50,
 ):
-    cosmology_parameters = __prep_cosmology_parameters(catalog.cosmology)
+    cosmology_parameters = utils.prep_cosmology_parameters(catalog.cosmology)
     catalog = catalog.evaluate(
-        age_at_z_,
+        utils.age_at_z_,
         vectorize=True,
         cosmology=cosmology_parameters,
         format="jax",
     )
 
-    batches = __split_central_indices(catalog, batch_size)
+    batches = utils.split_central_indices(catalog, batch_size)
 
     chunked_output = []
     for row_batch in batches:
